pyproject.toml
[tool.poetry]
name = "wpp-send-messages"
version = "0.1.0"
description = ""
authors = ["Evandro Paulo Folletto <evandrofolletto@gmail.com>"]
readme = "README.md"
packages = [{include = "wpp_send_messages"}]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.22.0"
pandas = "^2.2.2"
google-api-python-client = "^2.136.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
